+++ src/Freee.Accounting/Models/ExpenseApplicationUpdateParamsExpenseApplicationLines.cs
@@ -39,7 +39,7 @@ namespace Freee.Accounting.Models
         /// <param name="id">経費申請の項目行ID: 既存項目行を更新する場合に指定します。IDを指定しない項目行は、新規行として扱われ追加されます。また、expense_application_linesに含まれない既存の項目行は削除されます。更新後も残したい行は、必ず経費申請の項目行IDを指定してexpense_application_linesに含めてください。.</param>
         /// <param name="receiptId">証憑ID.</param>
         /// <param name="transactionDate">日付 (yyyy-mm-dd).</param>
-        public ExpenseApplicationUpdateParamsExpenseApplicationLines(int amount = default(int), string description = default(string), int expenseApplicationLineTemplateId = default(int), int id = default(int), int receiptId = default(int), string transactionDate = default(string))
+        public ExpenseApplicationUpdateParamsExpenseApplicationLines(int amount = default(int), string description = default(string), int expenseApplicationLineTemplateId = default(int), long id = default(long), int receiptId = default(int), string transactionDate = default(string))
         {
             this.Amount = amount;
             this.Description = description;
@@ -75,7 +75,7 @@ namespace Freee.Accounting.Models
         /// </summary>
         /// <value>経費申請の項目行ID: 既存項目行を更新する場合に指定します。IDを指定しない項目行は、新規行として扱われ追加されます。また、expense_application_linesに含まれない既存の項目行は削除されます。更新後も残したい行は、必ず経費申請の項目行IDを指定してexpense_application_linesに含めてください。</value>
         [DataMember(Name = "id", EmitDefaultValue = false)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 証憑ID
